+++ bevy_asset_loader/examples/failure_state.rs
@@ -35,7 +35,7 @@ fn fail() {
 fn ok(mut quit: EventWriter<AppExit>) {
     info!("As expected, bevy_asset_loader switched to the failure state");
     info!("Quitting the application...");
-    quit.send(AppExit);
+    quit.send(AppExit::Success);
 }
 
 fn timeout(time: Res<Time>) {