+++ bevy_asset_loader/src/loading_state/config.rs
@@ -13,7 +13,8 @@ use crate::loading_state::{
 use bevy::app::App;
 use bevy::asset::Asset;
 use bevy::ecs::schedule::SystemConfigs;
-use bevy::prelude::{default, FromWorld, IntoSystemConfigs, Resource, States};
+use bevy::prelude::{default, FromWorld, IntoSystemConfigs, Resource};
+use bevy::state::state::FreelyMutableState;
 use bevy::utils::HashMap;
 use std::any::TypeId;
 
@@ -67,7 +68,7 @@ pub trait ConfigureLoadingState {
 ///             .continue_to_state(GameState::Menu)
 ///         )
 ///         .configure_loading_state(LoadingStateConfig::new(GameState::Loading).load_collection::<AudioAssets>())
-/// #       .set_runner(|mut app| app.update())
+/// #       .set_runner(|mut app| {app.update(); AppExit::Success})
 ///         .run();
 /// # }
 ///
@@ -85,7 +86,7 @@ pub trait ConfigureLoadingState {
 /// #     plop: Handle<AudioSource>
 /// # }
 /// ```
-pub struct LoadingStateConfig<S: States> {
+pub struct LoadingStateConfig<S: FreelyMutableState> {
     state: S,
 
     on_enter_loading_assets: Vec<SystemConfigs>,
@@ -96,7 +97,7 @@ pub struct LoadingStateConfig<S: States> {
     dynamic_assets: HashMap<TypeId, Vec<String>>,
 }
 
-impl<S: States> LoadingStateConfig<S> {
+impl<S: FreelyMutableState> LoadingStateConfig<S> {
     /// Create a new configuration for the given loading state
     pub fn new(state: S) -> Self {
         Self {
@@ -144,7 +145,7 @@ impl<S: States> LoadingStateConfig<S> {
             );
         }
         let mut dynamic_assets = app
-            .world
+            .world_mut()
             .get_resource_mut::<DynamicAssetCollections<S>>()
             .unwrap_or_else(|| {
                 panic!("Failed to get the DynamicAssetCollections resource for the loading state. Are you trying to configure a loading state before it was added to the bevy App?")
@@ -155,7 +156,7 @@ impl<S: States> LoadingStateConfig<S> {
     }
 }
 
-impl<S: States> ConfigureLoadingState for LoadingStateConfig<S> {
+impl<S: FreelyMutableState> ConfigureLoadingState for LoadingStateConfig<S> {
     fn load_collection<A: AssetCollection>(mut self) -> Self {
         self.on_enter_loading_assets
             .push(start_loading_collection::<S, A>.into_configs());