+++ bevy_asset_loader/src/standard_dynamic_asset.rs
@@ -1,14 +1,13 @@
 use crate::dynamic_asset::{DynamicAsset, DynamicAssetType};
 use crate::dynamic_asset::{DynamicAssetCollection, DynamicAssets};
 use bevy::asset::{Asset, AssetServer, Assets, LoadedFolder, UntypedHandle};
-use bevy::ecs::system::Command;
-use bevy::ecs::world::World;
+use bevy::ecs::world::{Command, World};
 use bevy::reflect::TypePath;
 use bevy::utils::HashMap;
 use serde::{Deserialize, Serialize};
 
 #[cfg(feature = "2d")]
-use bevy::math::Vec2;
+use bevy::math::UVec2;
 #[cfg(feature = "3d")]
 use bevy::pbr::StandardMaterial;
 #[cfg(feature = "2d")]
@@ -60,25 +59,25 @@ pub enum StandardDynamicAsset {
     #[cfg(feature = "2d")]
     TextureAtlasLayout {
         /// The image width in pixels
-        tile_size_x: f32,
+        tile_size_x: u32,
         /// The image height in pixels
-        tile_size_y: f32,
+        tile_size_y: u32,
         /// Columns on the sprite sheet
-        columns: usize,
+        columns: u32,
         /// Rows on the sprite sheet
-        rows: usize,
+        rows: u32,
         /// Padding between columns in pixels
         #[serde(with = "optional", skip_serializing_if = "Option::is_none", default)]
-        padding_x: Option<f32>,
+        padding_x: Option<u32>,
         /// Padding between rows in pixels
         #[serde(with = "optional", skip_serializing_if = "Option::is_none", default)]
-        padding_y: Option<f32>,
+        padding_y: Option<u32>,
         /// Number of pixels offset of the first tile
         #[serde(with = "optional", skip_serializing_if = "Option::is_none", default)]
-        offset_x: Option<f32>,
+        offset_x: Option<u32>,
         /// Number of pixels offset of the first tile
         #[serde(with = "optional", skip_serializing_if = "Option::is_none", default)]
-        offset_y: Option<f32>,
+        offset_y: Option<u32>,
     },
 }
 
@@ -162,10 +161,11 @@ impl DynamicAsset for StandardDynamicAsset {
     }
 
     fn build(&self, world: &mut World) -> Result<DynamicAssetType, anyhow::Error> {
-        let cell = world.cell();
-        let asset_server = cell
-            .get_resource::<AssetServer>()
-            .expect("Cannot get AssetServer");
+        let cell = world.as_unsafe_world_cell();
+        let asset_server = unsafe {
+            cell.get_resource::<AssetServer>()
+                .expect("Cannot get AssetServer")
+        };
         match self {
             StandardDynamicAsset::File { path } => Ok(DynamicAssetType::Single(
                 asset_server.get_handle_untyped(path).unwrap(),
@@ -174,9 +174,10 @@ impl DynamicAsset for StandardDynamicAsset {
             StandardDynamicAsset::Image { path, sampler } => {
                 let mut handle = asset_server.load(path);
                 if let Some(sampler) = sampler {
-                    let mut images = cell
-                        .get_resource_mut::<Assets<Image>>()
-                        .expect("Cannot get resource Assets<Image>");
+                    let mut images = unsafe {
+                        cell.get_resource_mut::<Assets<Image>>()
+                            .expect("Cannot get resource Assets<Image>")
+                    };
                     Self::update_image_sampler(&mut handle, &mut images, sampler);
                 }
 
@@ -184,9 +185,10 @@ impl DynamicAsset for StandardDynamicAsset {
             }
             #[cfg(feature = "3d")]
             StandardDynamicAsset::StandardMaterial { path } => {
-                let mut materials = cell
-                    .get_resource_mut::<Assets<StandardMaterial>>()
-                    .expect("Cannot get resource Assets<StandardMaterial>");
+                let mut materials = unsafe {
+                    cell.get_resource_mut::<Assets<StandardMaterial>>()
+                        .expect("Cannot get resource Assets<StandardMaterial>")
+                };
                 let handle = materials
                     .add(StandardMaterial::from(
                         asset_server.get_handle::<Image>(path).unwrap(),
@@ -206,28 +208,30 @@ impl DynamicAsset for StandardDynamicAsset {
                 offset_x,
                 offset_y,
             } => {
-                let mut atlases = cell
-                    .get_resource_mut::<Assets<TextureAtlasLayout>>()
-                    .expect("Cannot get resource Assets<TextureAtlasLayout>");
+                let mut atlases = unsafe {
+                    cell.get_resource_mut::<Assets<TextureAtlasLayout>>()
+                        .expect("Cannot get resource Assets<TextureAtlasLayout>")
+                };
                 let texture_atlas_handle = atlases
                     .add(TextureAtlasLayout::from_grid(
-                        Vec2::new(*tile_size_x, *tile_size_y),
+                        UVec2::new(*tile_size_x, *tile_size_y),
                         *columns,
                         *rows,
-                        Some(Vec2::new(padding_x.unwrap_or(0.), padding_y.unwrap_or(0.))),
-                        Some(Vec2::new(offset_x.unwrap_or(0.), offset_y.unwrap_or(0.))),
+                        Some(UVec2::new(padding_x.unwrap_or(0), padding_y.unwrap_or(0))),
+                        Some(UVec2::new(offset_x.unwrap_or(0), offset_y.unwrap_or(0))),
                     ))
                     .untyped();
 
                 Ok(DynamicAssetType::Single(texture_atlas_handle))
             }
             StandardDynamicAsset::Folder { path } => {
-                let folders = cell
-                    .get_resource_mut::<Assets<LoadedFolder>>()
-                    .expect("Cannot get resource Assets<LoadedFolder>");
+                let folders = unsafe {
+                    cell.get_resource_mut::<Assets<LoadedFolder>>()
+                        .expect("Cannot get resource Assets<LoadedFolder>")
+                };
                 Ok(DynamicAssetType::Collection(
                     folders
-                        .get(asset_server.get_handle(path).unwrap())
+                        .get(&asset_server.get_handle(path).unwrap())
                         .unwrap()
                         .handles
                         .to_vec(),