+++ bevy_asset_loader/tests/can_run_without_next_state.rs
@@ -41,7 +41,7 @@ fn expect(
 ) {
     if collection.is_some() {
         if test_state.wait_frames_after_load == 0 {
-            exit.send(AppExit);
+            exit.send(AppExit::Success);
             return;
         }
         test_state.wait_frames_after_load -= 1;