+++ bevy_asset_loader/tests/continues_to_failure_state.rs
@@ -28,7 +28,7 @@ fn fail() {
 }
 
 fn exit(mut exit: EventWriter<AppExit>) {
-    exit.send(AppExit);
+    exit.send(AppExit::Success);
 }
 
 fn timeout(time: Res<Time>) {