+++ bevy_asset_loader/tests/reload_dynamic_asset_files.rs
@@ -84,5 +84,5 @@ fn timeout(time: Res<Time>) {
 
 fn quit(mut exit: EventWriter<AppExit>) {
     info!("Everything fine, quitting the app");
-    exit.send(AppExit);
+    exit.send(AppExit::Success);
 }