+++ bevy_asset_loader/tests/same_collection_multiple_times.rs
@@ -55,5 +55,5 @@ fn timeout(time: Res<Time>) {
 
 fn quit(mut exit: EventWriter<AppExit>) {
     info!("Everything fine, quitting the app");
-    exit.send(AppExit);
+    exit.send(AppExit::Success);
 }