+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- all texture atlas layout parameters are integers (u32) now
+
 ## v0.20.2 - 18.05.2024
 - Fix texture atlas layout in asset collections (see [#212](https://github.com/NiklasEi/bevy_asset_loader/pull/212))